fix page size issue
